final upload for Variadic functions

+++ 0x10-variadic_functions/3-print_all.c
@@ -67,7 +67,7 @@ void print_all(const char *const format, ...)
 				flag++;
 				break;
 			case 'i':
-				print_num(va_arg(args, double));
+				print_num(va_arg(args, int));
 				flag++;
 				break;
 			case 'f':
